+++ apps/ysceneview/ysceneview.cpp
@@ -31,8 +31,8 @@
 #include <yocto/yocto_image.h>
 #include <yocto/yocto_sceneio.h>
 #include <yocto/yocto_shape.h>
+#include <yocto_gui/yocto_draw.h>
 #include <yocto_gui/yocto_imgui.h>
-#include <yocto_gui/yocto_opengl.h>
 using namespace yocto;
 
 #include <atomic>
